Create tiles.roman.svg
I use autotable to train people at mahjong, but for those who can't read chinese characters it's very challenging. Here I've added numerals to the manzu tiles and the wind tiles. Unlike commercially available sets I try to avoid romanizing where it's not super helpful, to try and preserve the natural beauty of the tiles.

I am still trying to figure out how to make this a game option, but posting this here for initial feedback.